Add message cloning and rendering
This commit adds the ability to clone message sources and destinations
as well as cleaning up some behavior where the application would not
exit when Ctrl-C was sent from the terminal.

The exit behavior currently shows some errors, but does exit. It is
worth looking at how to get this exit behavior to act in a clean
manner in the future.